validate moved to frame.py
